transmitter: add an incomplete plugin (output and pointer)
Enabled with ./configure --enable-surface-remoting

Draft the public API for binding and unbinding a surface for remoting,
and another for ivi-shell specifics.

Add the plumbing needed to relay a configure event from the network to
the shell, to be forwarded to the application.

Implement support for (fake) connection object tracking.

Create a weston_output for a (fake) remote output. Force the
sync-to-output on remoted surfaces. Use timers to fake wl_surface.enter
and frame callback events. Also Presentation feedback is poorly faked.

Hooking to weston_surface happens in two ways: surface_configure in
Transmitter plugin API is used to relay the dx,dy from
wl_surface.attach, and weston_surface::apply_state_signal relays state
updates.

Create a fake seat with pointer:

- Add an internal pointer API that will be called by the networking code
  for incoming input events. The API should match 1:1 to the events.
  These functions will reimplement the parts of src/input.c to send out
  the events to Wayland clients.

- Implement pointer enter/leave of the internal pointer API. Code is
  mimicked from weston_pointer_set_focus(), but without all the logic
  needed to determine the right actions.

- Implement pointer frame of the internal pointer API.

- Add a fake pointer input generator. It only make the pointer enter the
  surface remoted the most recently, and moves the pointer in a circle.

All ways of getting pointer focus should be covered. Serial tracking is
the same as in Weston core. Seat maintains the pointer focus surface.
Focus_client is managed while Weston core never sees any focus.
Transmitter surface gained a destroy signal.

Surface destruction relies on forwarding the destroy to the remote,
which will then reply with appropriate pointer.leave events. This is not
implemented by the input faking code.

Transmitter is expected to open all inputs available on the remote and
multiplex those to local clients as clients subscribe.

Notes about the design and guidelines for finishing the implementation
are included, and all the mockup code is annotated as "fake" so it is
easy to spot and replace with a proper network implementation.

There are listener-based APIs for getting both remote connection status
changes, and per-surface stream status changes.

Signed-off-by: Pekka Paalanen <pekka.paalanen@collabora.co.uk>